# stm32: Enable link-time optimisation as a build option