--- a/ports/stm32/boards/NUCLEO_L073RZ/mpconfigboard.mk
+++ b/ports/stm32/boards/NUCLEO_L073RZ/mpconfigboard.mk
@@ -8,3 +8,6 @@ MICROPY_VFS_FAT = 0
 
 # Don't include default frozen modules because MCU is tight on flash space
 FROZEN_MANIFEST ?=
+
+# LTO reduces final binary size, may be slower to build depending on gcc version and hardware
+LTO ?= 1