--- a/tests/run-perfbench.py
+++ b/tests/run-perfbench.py
@@ -260,7 +260,9 @@ def main():
     )
     cmd_parser.add_argument("--via-mpy", action="store_true", help="compile code to .mpy first")
     cmd_parser.add_argument("--mpy-cross-flags", default="", help="flags to pass to mpy-cross")
-    cmd_parser.add_argument("N", nargs=1, help="N parameter (approximate target CPU frequency)")
+    cmd_parser.add_argument(
+        "N", nargs=1, help="N parameter (approximate target CPU frequency in MHz)"
+    )
     cmd_parser.add_argument("M", nargs=1, help="M parameter (approximate target heap in kbytes)")
     cmd_parser.add_argument("files", nargs="*", help="input test files")
     args = cmd_parser.parse_args()