… tests/perf_bench/core_import_mpy_multi.py
… tests/perf_bench/core_qstr.py