Adds capability to do `make LTO=1` for stm32 port. Currently disabled by default. Until now, nrf is the only MicroPython port with LTO support.

Goal is improving support on very small STM32 targets like the STM32L072 (20KB RAM, 192KB flash, 32MHz CPU). Every byte counts!

## Build Performance

Building test firmwares for B_L072Z_LRWAN1 and PYBV11. GCC is `arm-none-eabi-gcc (Arch Repository) 12.1.0`. Build system is a Ryzen 5 3600 (6 Cores 12 Threads). 

Build method is `make clean BOARD=B_L072Z_LRWAN1 && time make -j BOARD=B_L072Z_LRWAN1 LTO=1`. No ccache. Best of 3 runs for build time.

| Board          | Options | Build Time                                      | text segment size | difference   |
|----------------|---------|-------------------------------------------------|-------------------|--------------|
| B_L072Z_LRWAN1 | LTO=0   | 54.12s user 8.99s system 735% cpu 8.583 total   | 193420            |              |
|                | LTO=1   | 49.97s user 8.95s system 666% cpu 8.839 total   | 188040            | -5380 (2.8%) |
| PYBV11         | LTO=0   | 70.53s user 10.51s system 755% cpu 10.731 total | 369228            |              |
|                | LTO=1   | 69.05s user 10.60s system 735% cpu 10.834 total | 360812            | -8416 (2.3%) |

Difference is pretty minor, but measurable (and 5KB on a board that currently has <3KB free in the default image is not too shabby!)

Surprisingly, on my system the build times (CPU time and wall time) seemed basically the same with or without LTO, and LTO is actually a little faster in CPU time. This might be different with older gcc versions or more constrained hardware.

## Runtime Performance

Testing on Pyboard V1.1, run as `./run-perfbench.py -p -d /dev/ttyACM0 168 100`

```
> ./run-perfbench.py -p -d /dev/ttyACM0 -s pyboardv11-no-lto.txt pyboardv11-lto.txt
diff of scores (higher is better)
N=168 M=100                pyboardv11-no-lto.txt -> pyboardv11-lto.txt         diff      diff% (error%)
bm_chaos.py                    357.38 ->     352.89 :      -4.49 =  -1.256% (+/-0.01%)
bm_fannkuch.py                  77.25 ->      77.52 :      +0.27 =  +0.350% (+/-0.01%)
bm_fft.py                     2524.17 ->    2516.82 :      -7.35 =  -0.291% (+/-0.00%)
bm_float.py                   5877.62 ->    5748.99 :    -128.63 =  -2.188% (+/-0.02%)
bm_hexiom.py                    42.25 ->      42.22 :      -0.03 =  -0.071% (+/-0.00%)
bm_nqueens.py                 4372.41 ->    4407.51 :     +35.10 =  +0.803% (+/-0.00%)
bm_pidigits.py                 635.78 ->     633.96 :      -1.82 =  -0.286% (+/-0.17%)
core_import_mpy_multi.py       477.10 ->     477.72 :      +0.62 =  +0.130% (+/-0.01%)
core_import_mpy_single.py       58.76 ->      58.74 :      -0.02 =  -0.034% (+/-0.00%)
core_qstr.py                    63.38 ->      63.11 :      -0.27 =  -0.426% (+/-0.00%)
core_yield_from.py             359.39 ->     357.57 :      -1.82 =  -0.506% (+/-0.00%)
misc_aes.py                    408.89 ->     397.27 :     -11.62 =  -2.842% (+/-0.00%)
misc_mandel.py                3433.68 ->    3375.74 :     -57.94 =  -1.687% (+/-0.01%)
misc_pystone.py               2310.39 ->    2265.40 :     -44.99 =  -1.947% (+/-0.00%)
misc_raytrace.py               369.97 ->     367.60 :      -2.37 =  -0.641% (+/-0.01%)
viper_call0.py                 590.87 ->     605.92 :     +15.05 =  +2.547% (+/-0.00%)
viper_call1a.py                563.23 ->     576.78 :     +13.55 =  +2.406% (+/-0.00%)
viper_call1b.py                443.97 ->     452.46 :      +8.49 =  +1.912% (+/-0.01%)
viper_call1c.py                448.74 ->     457.39 :      +8.65 =  +1.928% (+/-0.00%)
viper_call2a.py                548.53 ->     561.37 :     +12.84 =  +2.341% (+/-0.00%)
viper_call2b.py                383.19 ->     389.49 :      +6.30 =  +1.644% (+/-0.01%)
```

A few percent slower in some tests (standout is misc_aes.py at -2.8%), but faster in some others (all viper calls got about 2% faster). Can't explain why LTO would make some operations slower...

On STM32L072 (B-L072Z-LRWAN board), run as `./run-perfbench.py -p -d /dev/ttyACM0 32 10`
                                                                                                           
```
> ./run-perfbench.py -p -d /dev/ttyACM0 -s stm32l072-no-lto.txt stm32l072-lto.txt
diff of scores (higher is better)
N=32 M=10                  stm32l072-no-lto.txt -> stm32l072-lto.txt         diff      diff% (error%)
bm_fannkuch.py                 873.71 ->     874.48 :      +0.77 =  +0.088% (+/-0.08%)
bm_nqueens.py                  333.55 ->     334.51 :      +0.96 =  +0.288% (+/-0.00%)
bm_pidigits.py                 133.18 ->     134.70 :      +1.52 =  +1.141% (+/-0.09%)
core_import_mpy_multi.py        57.39 ->      57.61 :      +0.22 =  +0.383% (+/-0.06%)
core_qstr.py                    11.07 ->      11.09 :      +0.02 =  +0.181% (+/-0.00%)
```

Most tests don't run in this resource footprint. LTO slightly faster, but not really.

## Other tradeoffs for enabling LTO

* Source-level native code debugging is more of a challenge, as execution flow can jump unexpectedly between source files.
* Undefined Behaviour that isn't an issue with LTO=0 may become exposed with LTO=1.
* Marginally higher possibility of hitting compiler bugs, although LTO is pretty mature these days...