@dpgeorge 2022-06-23
Thanks, this is nice!

I'd suggest two things:
- Decide on a name for the makefile-level `LTO` variable, so that we can use this consistently across all ports.  Is `LTO` unique enough that it won't clash with anything else?  It could be `USE_LTO` or `MICROPY_LTO`?
- Enable LTO on small stm32 boards, I'd say F0, L0, G0 and L432KC