@jimmo 2022-06-23
> Marginally higher possibility of hitting compiler bugs, although LTO is pretty mature these days...

I don't know anything about the history of LTO and compiler bugs, but if this is a thing then do we need to consider turning it off by default below certain compiler versions? I think people are still regularly using 8.x and 9.x for example.

@tannewt @dhalbert @jepler CircuitPython uses LTO right? I saw in https://github.com/adafruit/circuitpython/commit/9d10a3da66cd799b659457ca9785be1a33105dc6 that it looks like you're seeing problems on GCC 10.x ?