@dhalbert 2022-06-23
> @tannewt @dhalbert @jepler CircuitPython uses LTO right? I saw in [adafruit@9d10a3d](https://github.com/adafruit/circuitpython/commit/9d10a3da66cd799b659457ca9785be1a33105dc6) that it looks like you're seeing problems on GCC 10.x ?

We have been using LTO for years, and in general are quite happy with it, though it does make for slower builds. It can make a huge difference in code size. We have used it on the SAMD and nRF ports, but have not yet tried it on STM or Espressif (I think

We did some tricky things in the commit you mentioned to improve compilation of the translation strings when using LTO, but there is not an inherent problem with gcc10: see https://github.com/adafruit/circuitpython/pull/6440.  And we used it on versions below gcc10 without particular difficulty. Sometimes you have to make sure certain sections, such as the interrupt vector, don't disappear, but the problems have all been solvable.

@tannewt and @jepler may have further comments as well.