@dhalbert 2022-06-23
Also see https://github.com/adafruit/circuitpython/pull/6436, which did some tuning of an LTO option.