@chrismas9 2022-06-23
`Difference is pretty minor`

On the F091 port the filesystem size is 20K with 1700 bytes spare. This will increase it to 26K. That's not minor, it's massive. All filesystem sizes on small chips should be increased if this is implemented.