@tannewt 2022-06-23
Generally LTO has been great for the smallest builds.

Another caveat to know is that it's not great for multi-tiered memory systems because it may merge code across memory regions. (For example merging RAM-only code into flash stored code.) Our small micros don't have this memory setup though.