@projectgus 2022-06-24
Thank you @tannewt and @dhalbert for the context, that's really useful! It's interesting you see size improvements with `-flto-partition=one` over `balanced`, I'm pretty sure I didn't notice a size change on the stm32 builds (but this could be for any number of reasons). @dhalbert do you happen to know what gcc version you used for those benchmarks?

If that's the case then it's probably worth parameterising the partition method in the variable, same as you've done for CircuitPython.