@projectgus 2022-06-24
Fairly unexpected finding. LTO builds with default "balanced" partitioning are *not significantly slower* on gcc 12.1 than LTO=0, even on pretty low end hardware.

PYBV11 builds on both my 2016 Dell laptop (i7-6500U 2 Cores 4 Threads)  and my desktop Ryzen 5 3600 (6 Cores 12 Threads) are both within about 2% total wall time of LTO=0 vs LTO=1, which I think is within the margin of error.

Building for NUCLEO_L073RZ on my desktop the LTO build is 0.7 seconds slower, which is 8% slower.