@projectgus 2022-06-24
@dpgeorge Thanks for the useful feedback!

>   Decide on a name for the makefile-level LTO variable, so that we can use this consistently across all ports. Is LTO unique enough that it won't clash with anything else? It could be USE_LTO or MICROPY_LTO?

The nrf port is using `LTO` already. Although it's a short acronym, I think it's only ever commonly used as an environment variable for this purpose so if I'd suggest keeping it if you're comfortable with that (in the same way names like CFLAGS are treated as generic rather than scoped to a project).

>    Enable LTO on small stm32 boards, I'd say F0, L0, G0 and L432KC

Ack, will do!

> I don't know anything about the history of LTO and compiler bugs, but if this is a thing then do we need to consider turning it off by default below certain compiler versions? I think people are still regularly using 8.x and 9.x for example.

I think it's probably OK to keep it on regardless of version. I think I've been using LTO since gcc 4.x and I've never run into a bug that was caused by the compiler due to LTO. I have definitely run into bugs that were only exposed by enabling LTO, but those were actual bugs or UB. But it will be a thing to keep in mind when reproducing a bug, that it's possible it will appear or disappear depending on the LTO setting and gcc version...