@projectgus 2022-06-24
@dpgeorge @jimmo (& others)

One thing that does happen is that if you do a "dirty" config change from LTO=0 to LTO=1 builds without also doing a clean build, you can see some pretty scary looking linker errors (that go away with a clean build). If this is likely to be a problem then an easy way around it is to put a suffix on all the LTO object files (like name-lto.o vs name.o).

The other approach about it would be to add a note in the docs, though. Which made me wonder about build docs. Is it worth adding a section about LTO this in the port Makefile, or even a generic "Options when building MicroPython" section in the docs themselves? Happy to write such a section if so. Apologies if there is one already, but I missed it.