@projectgus 2022-06-24
Have rebased, and also updated with some new perfbench runs on Pyboard and STM32L072.