@dpgeorge 2022-06-24
> One thing that does happen is that if you do a "dirty" config change from LTO=0 to LTO=1 builds without also doing a clean build, you can see some pretty scary looking linker errors (that go away with a clean build).

Yeah, this happens with lots of other things as well, basically changing any Makefile-level option.  Eg `MICROPY_PY_BLUETOOTH` on PYBD_SF6.  I'd say we can let that be as it is for now, and defer it to another time (maybe never) to fix in a comprehensive way.