@dpgeorge 2022-06-24
> Which made me wonder about build docs. Is it worth adding a section about LTO this in the port Makefile, or even a generic "Options when building MicroPython" section in the docs themselves?

Adding a not to stm32's Makefile would be worth doing (nrf has a similar note).  For generic docs it'd possibly go in `docs/develop/gettingstarted.rst`.