@dhylands 2022-06-26
> Yeah, this happens with lots of other things as well, basically changing any Makefile-level option. Eg `MICROPY_PY_BLUETOOTH` on PYBD_SF6. I'd say we can let that be as it is for now, and defer it to another time (maybe never) to fix in a comprehensive way.

One simple way is to make all of the objects depend on all of the Makefiles (available in $(MAKEFILE_LIST).