@projectgus 2022-06-27
> One simple way is to make all of the objects depend on all of the Makefiles (available in $(MAKEFILE_LIST).

@dhylands This is a good practice (albeit not without footguns), but it doesn't solve this use case of overriding some variables on the make command line. I don't know of a good way around that one.