@H-Grobben 2022-07-26
Situation: I need some Fuses set in the processor (to be exact, disable the Boot0 pin functionality because the device would often go into bootloader mode). I added the Fuses to a .c file in the board folder, with a .ld file with the locations of the fuses. When I build without LTO, the fuses are found in the Hex file. If I use LTO=1 during build, the fuses are absent in the hex file. Is this expected behavior?