@jimmo 2022-07-26
@H-Grobben you need to set the `used` attribute, e.g.

`const uint32_t optionbytes1  __attribute__(( section(".optionbytes1"), used )) = 0xFBEFF8AA;`

(It's not enough to do this via `KEEP` in the linker script)